SypanseSquad/requirements.txt
pandas
kafka-python
pyspark==3.5.0
